+++ src/PlanoPersonalizado.cpp
@@ -1,9 +1,11 @@
 #include "../header/PlanoPersonalizado.h"
 
-PlanoPersonalizado::PlanoPersonalizado(const std::string &descricao, double valor) : Plano(descricao, valor) {}
+PlanoPersonalizado::PlanoPersonalizado(const std::string &descricao,
+                                       double valor)
+    : Plano(descricao, valor) {}
 
-int PlanoPersonalizado::getId() const { return id; }
+size_t PlanoPersonalizado::getId() const { return id; }
 
-void PlanoPersonalizado::setId(int i) { id = i; }
+void PlanoPersonalizado::setId(size_t i) { id = i; }
 
 double PlanoPersonalizado::calcularValor() const { return Plano::valor; }
